feat(): Separação de fução de processamento
