+++ tests/run_test_dps_sensor.py
@@ -4,11 +4,13 @@ import pandas as pd
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from sensors.dps_sensor import SensorDPS
+from process.dps_processor import DPSProcessor
 
 if __name__ == "__main__":
-    sensor = SensorDPS(qtdGerada=96)
-    df = sensor.generate_data()
+    qtd_dados = 96
+    processor = DPSProcessor(qtdGerada=qtd_dados)
+
+    df = processor.generate_dataframe()
 
     output_dir = os.path.join(os.path.dirname(__file__), "output")
     os.makedirs(output_dir, exist_ok=True)
